backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
xgboost
requests
python-dotenv
